+++ HISTORY.rst
@@ -25,6 +25,8 @@ latest (unreleased)
 
 **Bugfixes**
 
+* BSSFL-556: Propagate lot from sale line through the 3 delivery steps
+
 **Build**
 
 **Documentation**